introduce proxy_ext
