+++ engine_tiflash/src/engine.rs
@@ -8,10 +8,7 @@ use std::{
     fs,
     ops::Deref,
     path::Path,
-    sync::{
-        atomic::{AtomicIsize, Ordering},
-        Arc,
-    },
+    sync::{atomic::AtomicIsize, Arc},
 };
 
 use engine_rocks::{RocksDbVector, RocksEngineIterator, RocksSnapshot};
@@ -25,31 +22,8 @@ use crate::{
     proxy_utils::{engine_ext::*, EngineStoreHub},
     r2e,
     util::get_cf_handle,
+    ProxyEngineExt,
 };
-
-// This struct should be safe to copy.
-#[derive(Clone)]
-pub struct ProxyEngineExt {
-    pub engine_store_server_helper: isize,
-    pub pool_capacity: usize,
-    pub pending_applies_count: Arc<AtomicIsize>,
-    pub engine_store_hub: Option<Arc<dyn EngineStoreHub + Send + Sync>>,
-    pub config_set: Option<Arc<crate::ProxyConfigSet>>,
-    pub cached_region_info_manager: Option<Arc<crate::CachedRegionInfoManager>>,
-}
-
-impl Default for ProxyEngineExt {
-    fn default() -> Self {
-        ProxyEngineExt {
-            engine_store_server_helper: 0,
-            pool_capacity: 0,
-            pending_applies_count: Arc::new(AtomicIsize::new(0)),
-            engine_store_hub: None,
-            config_set: None,
-            cached_region_info_manager: None,
-        }
-    }
-}
 
 #[derive(Clone)]
 pub struct RocksEngine {
@@ -165,51 +139,9 @@ impl KvEngine for RocksEngine {
         self.rocks.bad_downcast()
     }
 
-    // The whole point is:
-    // 1. When `handle_pending_applies` is called by `on_timeout`, we can handle at
-    // least one.
-    // 2. When `handle_pending_applies` is called when we receive a
-    // new task, or when `handle_pending_applies` need to handle multiple snapshots.
-    // We need to compare to what's in queue.
     fn can_apply_snapshot(&self, is_timeout: bool, new_batch: bool, region_id: u64) -> bool {
-        fail::fail_point!("on_can_apply_snapshot", |e| e
-            .unwrap()
-            .parse::<bool>()
-            .unwrap());
-        if let Some(s) = self.proxy_ext.config_set.as_ref() {
-            if s.engine_store.enable_fast_add_peer {
-                // TODO Return true if this is an empty snapshot.
-                // We need to test if the region is still in fast add peer mode.
-                let result = self
-                    .proxy_ext
-                    .cached_region_info_manager
-                    .as_ref()
-                    .expect("expect cached_region_info_manager")
-                    .get_inited_or_fallback(region_id);
-                match result {
-                    Some(true) => {
-                        // Do nothing.
-                        tikv_util::debug!("can_apply_snapshot no fast path. do normal checking";
-                            "region_id" => region_id,
-                        );
-                    }
-                    None | Some(false) => {
-                        // Otherwise, try fast path.
-                        return true;
-                    }
-                };
-            }
-        }
-        // is called after calling observer's pre_handle_snapshot
-        let in_queue = self.proxy_ext.pending_applies_count.load(Ordering::SeqCst);
-        if is_timeout && new_batch {
-            // If queue is full, we should begin to handle
-            true
-        } else {
-            // Otherwise, we wait until the queue is full.
-            // In order to batch more tasks.
-            in_queue > (self.proxy_ext.pool_capacity as isize)
-        }
+        self.proxy_ext
+            .can_apply_snapshot(is_timeout, new_batch, region_id)
     }
 }
